refactoting

+++ src/bai_thuc_hanh3/PhepNhan_ThanhBao.java
@@ -1,7 +1,15 @@
+/**
+ * 
+ */
+/**
+ * @author Bao Phan
+ */
 package bai_thuc_hanh3;
-
 public class PhepNhan_ThanhBao {
- int nhan(int a, int b) {
+	// TODO ham nhan
+	// sua int thanh long
+ long nhan(int a, int b) {
+	 // ko cần biến c, cho trả về giá trị a*b
 	 return a*b;
  }
 }
